// frontend/src/types/contracts/EventContract.ts
export declare namespace Event {
  export type DataStruct = {
    price: PromiseOrValue<BigNumberish>;
    time_start: PromiseOrValue<BigNumberish>;
    time_end: PromiseOrValue<BigNumberish>;
    event_name: PromiseOrValue<string>;
    event_symbol: PromiseOrValue<string>;
    company_name: PromiseOrValue<string>;
    eventLocation: PromiseOrValue<string>;
    event_address: PromiseOrValue<string>;
    owner: PromiseOrValue<string>;
    isOfflineEvent: PromiseOrValue<boolean>;
  };

  export type DataStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    string,
    string,
    string,
    string,
    string,
    boolean
  ] & {
    price: BigNumber;
    time_start: BigNumber;
    time_end: BigNumber;
    event_name: string;
    event_symbol: string;
    company_name: string;
    eventLocation: string;
    event_address: string;
    owner: string;
    isOfflineEvent: boolean;
  };
}